refactor(ui-scripts): relocate lint command out of legacy lib/test/ folder
Move lint.ts from lib/test/ into lib/commands/, alongside the other yargs
command modules (bump.ts, server.ts, etc.) it is registered with in
commands/index.ts. The test/ directory is a leftover name from when
ui-scripts absorbed the separate ui-test package.
