+++ packages/ui-scripts/lib/commands/index.ts
@@ -30,7 +30,7 @@ import publish from './publish.ts'
 import publishPrivate from './publish-private.ts'
 import visualDiff from './visual-diff.ts'
 import transpileDiff from './transpile-diff.ts'
-import lint from '../test/lint.ts'
+import lint from './lint.ts'
 import bundle from '../build/webpack.ts'
 import clean from '../build/clean.ts'
 import build from '../build/babel.ts'
